Move the TokenChecker to the libs and move the source tests there
