+++ config.js.example
@@ -6,7 +6,7 @@ global.DEFAULT_LOG_LOCATION = '.';
 global.DEFAULT_RESULT_LOCATION = '/var/www/';
 global.DEFAULT_PORT = 3000;
 global.RESOURCES_WHITELIST = './whitelist.txt';
-global.TOKEN_ENDPOINT = 'https://www.w3.org/Web/publications/authorize';
+global.TOKEN_ENDPOINT = 'http://localhost:3001/authorize';
 global.SPEC_GENERATOR = 'http://localhost:3002/';
 global.USERNAME = process.env.USERNAME;
 global.PASSWORD = process.env.PASSWORD;
