+++ lib/orchestrator.js
@@ -11,7 +11,7 @@ var JsonHttpService = require('./json-http-service');
 var Publisher = require('./publisher');
 var SpecberusWrapper = require('../functions.js').SpecberusWrapper;
 var ThirdPartyChecker = require('./third-party-resources-checker');
-var TokenChecker = require('../functions.js').TokenChecker;
+var TokenChecker = require('./token-checker').TokenChecker;
 
 // Configuration file
 require('../config.js');
@@ -178,25 +178,17 @@ Orchestrator.prototype.runSpecberus = function (httpLocation) {
 Orchestrator.prototype.runTokenChecker = function (latestVersion, url, token) {
   return new Map({
     name: 'token-checker',
-    promise: TokenChecker.check(latestVersion, token)
-      .then(function (authReport) {
-        var simpleSource1 = authReport.source.replace(/^https:/i, 'http:');
-        var simpleSource2 = authReport.source.replace(/^http:/i, 'https:');
-        var matchSource = (
-          url.indexOf(simpleSource1) === 0 ||
-          url.indexOf(simpleSource2) === 0
-        );
-
-        if (authReport.authorized && matchSource) {
+    promise: TokenChecker.check(latestVersion, token, url)
+      .then(function (errors) {
+        if (errors.isEmpty()) {
           return new Map({
-            status: 'ok',
-            history: 'You are authorized to publish'
+            status: 'ok'
           });
         }
         else {
           return new Map({
             status: 'failure',
-            errors: List.of('Not authorized'),
+            errors: errors,
             history: 'You are not authorized to publish'
           });
         }
