+++ test/lib/tokenchecker.js
@@ -2,14 +2,7 @@
 'use strict';
 
 var express = require('express');
-var getMetadata = require('./utils').getMetadata;
 var tokenChecker = express();
-
-function getShortName(latestVersion) {
-  var str = latestVersion.substring(0, latestVersion.lastIndexOf('/'));
-
-  return str.substring(str.lastIndexOf('/') + 1);
-}
 
 tokenChecker.get('/authorize', function (req, res) {
   var uri = req.query.spec;
@@ -17,7 +10,7 @@ tokenChecker.get('/authorize', function (req, res) {
   var json = {
     token: token,
     spec: uri,
-    source: getMetadata(getShortName(uri)).editorsDraft,
+    source: 'http://localhost:3001/drafts/',
     authorized: true
   };
